#32 Move type PieceWasCaptured and PieceWasMoved to separate files
