+++ src/app/model/PieceWasCaptured.ts
@@ -1,0 +1,8 @@
+import { Piece } from './Piece';
+import { Square } from './Types';
+
+export type PieceWasCaptured = {
+  eventType: 'PieceWasCaptured';
+  piece: Piece;
+  onSquare: Square;
+};
