+++ src/app/model/PieceWasMoved.ts
@@ -1,0 +1,9 @@
+import { Piece } from './Piece';
+import { Square } from './Types';
+
+export type PieceWasMoved = {
+  eventType: 'PieceWasMoved';
+  piece: Piece;
+  from: Square;
+  to: Square;
+};
